# Replace the `XRef.cache` Array with a Map instead